Update README.md
This application let user scan images of the document/book, which the app converts them to pdf and allows for certain modifications like displaying a specific page, updating the last page and removing any specific page of the pdf.

To deploy this app to android we have used the Buildozer packaging tool whose .specification file has been added to the repository

The current path of the pdf and images are saved in '/storage/emulated/0/DCIM' which is the DCIM directory inside the Internal Shared Storage of the android phone. To change this path, please make changes in the line 25,30,34,43 and 50 to the path where you want to set.